// src/themes/generatedTokens.ts
const figmaTokens = {
  "Variables": {
    "Light": {
      "Theme": {
        "chip-background": {
          "$value": "#2f2b3d14"
        },
        "input-border": {
          "$value": "#2f2b3d38",
          "$description": "Border style for standard variant Text Field & Select"
        },
        "filled-input-bg": {
          "$value": "#2f2b3d0f"
        },
        "backdrop-overlay": {
          "$value": "#2f2b3d80",
          "$description": "Backdrop overlay style"
        },
        "outline-border": {
          "$value": "#2f2b3d3d"
        },
        "divider": {
          "$value": "#2f2b3d1f",
          "$description": "Divider fill color"
        },
        "action-focus": {
          "$value": "#2f2b3d1a",
          "$description": "Fill background for components in focus state (List, Table, etc)"
        },
        "action-disabled-bg": {
          "$value": "#2f2b3d29",
          "$description": "Fill background for components in disabled state (List, Table, etc)"
        },
        "action-disabled": {
          "$value": "#2f2b3d4d",
          "$description": "Content color for components in disabled state (Button, List, Table, etc)"
        },
        "action-selected": {
          "$value": "#2f2b3d14",
          "$description": "Fill background for components in selected state (List, Table, etc)"
        },
        "action-hover": {
          "$value": "#2f2b3d0f",
          "$description": "Fill background for components in hover state (List, Table, etc)"
        },
        "text-disabled": {
          "$value": "#2f2b3d66"
        },
        "text-subtitle": {
          "$value": "#2f2b3d8c",
          "$description": "Text primary"
        },
        "action-active": {
          "$value": "#2f2b3d99",
          "$description": "Fill color for components in active state (List, Table, etc)"
        },
        "text-secondary": {
          "$value": "#2f2b3db2",
          "$description": "Text secondary"
        },
        "text-primary": {
          "$value": "#2f2b3de5",
          "$description": "Text primary"
        },
        "$type": "color"
      },
// ...
      "Gap": {
        "gap-1": {
          "$value": "{Primitives.Light.Gap.gap-1}"
        },
        "gap-2": {
          "$value": "{Primitives.Light.Gap.gap-2}"
        },
        "gap-3": {
          "$value": "{Primitives.Light.Gap.gap-3}"
        },
        "gap-4": {
          "$value": "{Primitives.Light.Gap.gap-4}"
        },
        "gap-5": {
          "$value": "{Primitives.Light.Gap.gap-5}"
        },
        "gap-6": {
          "$value": "{Primitives.Light.Gap.gap-6}"
        },
        "gap-7": {
          "$value": "{Primitives.Light.Gap.gap-7}"
        },
        "gap-8": {
          "$value": "{Primitives.Light.Gap.gap-8}"
        },
        "gap-9": {
          "$value": "{Primitives.Light.Gap.gap-9}"
        },
        "gap-10": {
          "$value": "{Primitives.Light.Gap.gap-10}"
        },
        "gap-11": {
          "$value": "{Primitives.Light.Gap.gap-11}"
        },
        "gap-12": {
          "$value": "{Primitives.Light.Gap.gap-12}"
        },
        "gap-13": {
          "$value": "{Primitives.Light.Gap.gap-13}"
        },
        "gap-14": {
          "$value": "{Primitives.Light.Gap.gap-14}"
        },
        "gap-15": {
          "$value": "{Primitives.Light.Gap.gap-15}"
        },
        "gap-16": {
          "$value": "{Primitives.Light.Gap.gap-16}"
        },
        "$type": "dimension"
      },
      "Border-Radius": {
        "border-radius-xs": {
          "$value": "{Primitives.Light.Border-Radius.border-radius-xs}"
        },
        "border-radius-sm": {
          "$value": "{Primitives.Light.Border-Radius.border-radius-sm}"
        },
        "border-radius-md": {
          "$value": "{Primitives.Light.Border-Radius.border-radius-md}"
        },
        "border-radius-lg": {
          "$value": "{Primitives.Light.Border-Radius.border-radius-lg}"
        },
        "border-radius-xl": {
          "$value": "{Primitives.Light.Border-Radius.border-radius-xl}"
        },
        "border-round": {
          "$value": "{Primitives.Light.Border-Radius.border-round}"
        },
        "$type": "dimension"
      },
      "Misc": {
// ...
        "body-bg": {
          "$value": "#f8f7fa",
          "$type": "color"
        },
        "grey-light": {
          "$value": "#fafafa",
          "$type": "color"
        },
        "bg-white": {
          "$value": "#ffffff",
          "$type": "color"
        },
// ...
      }
    },
// ...
  },
  "Primitives": {
    "Light": {
// ...
      "Color": {
        "Primary": {
          "primary-50": {
            "$value": "#f5f3ff"
          },
          "primary-100": {
            "$value": "#e3e1fc"
          },
          "primary-200": {
            "$value": "#c7c2f9"
          },
          "primary-300": {
            "$value": "#aba4f6"
          },
          "primary-400": {
            "$value": "#8f85f3"
          },
          "primary-500": {
            "$value": "#7367f0"
          },
          "primary-600": {
            "$value": "#675dd8"
          },
          "primary-700": {
            "$value": "#6258cc"
          },
          "primary-800": {
            "$value": "#5c52c0"
          },
          "primary-900": {
            "$value": "#564db4"
          }
        },
        "Secondary": {
          "secondary-100": {
            "$value": "#e6e6e9"
          },
          "secondary-200": {
            "$value": "#cccdd3"
          },
          "secondary-300": {
            "$value": "#b3b5bc"
          },
          "secondary-400": {
            "$value": "#999ca6"
          },
          "secondary-500": {
            "$value": "#808390"
          },
          "secondary-600": {
            "$value": "#737682"
          },
          "secondary-700": {
            "$value": "#6d6f7a"
          },
          "secondary-800": {
            "$value": "#666973"
          },
          "secondary-900": {
            "$value": "#60626c"
          }
        },
        "Info": {
          "info-100": {
            "$value": "#ccf1f6"
          },
          "info-200": {
            "$value": "#99e3ed"
          },
          "info-300": {
            "$value": "#66d6e3"
          },
          "info-400": {
            "$value": "#33c8da"
          },
          "info-500": {
            "$value": "#00bad1"
          },
          "info-600": {
            "$value": "#00a7bc"
          },
          "info-700": {
            "$value": "#009eb2"
          },
          "info-800": {
            "$value": "#0095a7"
          },
          "info-900": {
            "$value": "#008c9d"
          }
        },
        "Success": {
          "success-100": {
            "$value": "#d4f4e2"
          },
          "success-200": {
            "$value": "#a9e9c5"
          },
          "success-300": {
            "$value": "#7edda9"
          },
          "success-400": {
            "$value": "#53d28c"
          },
          "success-500": {
            "$value": "#28c76f"
          },
          "success-600": {
            "$value": "#24b364"
          },
          "success-700": {
            "$value": "#22a95e"
          },
          "success-800": {
            "$value": "#209f59"
          },
          "success-900": {
            "$value": "#1e9553"
          }
        },
        "Warning": {
          "warning-100": {
            "$value": "#ffecd9"
          },
          "warning-200": {
            "$value": "#ffd9b4"
          },
          "warning-300": {
            "$value": "#ffc58e"
          },
          "warning-400": {
            "$value": "#ffb269"
          },
          "warning-500": {
            "$value": "#ff9f43"
          },
          "warning-600": {
            "$value": "#e68f3c"
          },
          "warning-700": {
            "$value": "#d98739"
          },
          "warning-800": {
            "$value": "#cc7f36"
          },
          "warning-900": {
            "$value": "#bf7732"
          }
        },
        "Error": {
          "error-100": {
            "$value": "#ffdbdc"
          },
          "error-200": {
            "$value": "#ffb7b9"
          },
          "error-300": {
            "$value": "#ff9497"
          },
          "error-400": {
            "$value": "#ff7074"
          },
          "error-500": {
            "$value": "#ff4c51"
          },
          "error-600": {
            "$value": "#e64449"
          },
          "error-700": {
            "$value": "#d94145"
          },
          "error-800": {
            "$value": "#cc3d41"
          },
          "error-900": {
            "$value": "#bf393d"
          }
        },
        "Gray": {
          "gray-100": {
            "$value": "#eaeaec"
          },
          "gray-200": {
            "$value": "#d5d4d8"
          },
          "gray-300": {
            "$value": "#c0bec5"
          },
          "gray-400": {
            "$value": "#aba8b1"
          },
          "gray-500": {
            "$value": "#97939e"
          },
          "gray-600": {
            "$value": "#827d8b"
          },
          "gray-700": {
            "$value": "#6d6777"
          },
          "gray-800": {
            "$value": "#585164"
          },
          "gray-900": {
            "$value": "#433c50"
          }
        },
        "$type": "color"
      },
// ...
      "Typography": {
        "font-family": {
          "$value": "'Inter', sans-serif",
          "$type": "fontFamily"
        },
        "font-size": {
          "12": {
            "$value": "12px"
          },
          "13": {
            "$value": "13px"
          },
          "15": {
            "$value": "15px"
          },
          "18": {
            "$value": "18px"
          },
          "24": {
            "$value": "24px"
          },
          "28": {
            "$value": "28px"
          },
          "38": {
            "$value": "38px"
          },
          "46": {
            "$value": "46px"
          },
          "$type": "dimension"
        },
        "font-weight": {
          "regular": {
            "$value": "400"
          },
          "medium": {
            "$value": "500"
          },
          "semibold": {
            "$value": "600"
          },
          "bold": {
            "$value": "700"
          },
          "$type": "fontWeight"
        },
        "line-height": {
          "14": {
            "$value": "14px"
          },
          "18": {
            "$value": "18px"
          },
          "20": {
            "$value": "20px"
          },
          "22": {
            "$value": "22px"
          },
          "28": {
            "$value": "28px"
          },
          "38": {
            "$value": "38px"
          },
          "42": {
            "$value": "42px"
          },
          "56": {
            "$value": "56px"
          },
          "68": {
            "$value": "68px"
          },
          "$type": "dimension"
        }
      }
    },
// ...
  }
};

export const themeKeys = {"Light":"light","Dark":"dark"} as const;